[EI-167] [web] refactor: Update metadata component imports
