+++ web/__tests__/components/numerical-guidance/metadata-dialog-menu.test.tsx
@@ -2,8 +2,8 @@ import { render, screen } from '@testing-library/react';
 import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
 import { resetMockDB } from '@/app/mocks/db.mock';
 import { resetAllStore } from '@/app/store/stores/reset-store';
-import MetdataList from '@/app/ui/components/numerical-guidance/molecule/metadata-list';
-import MetadataDialogMenu from '@/app/ui/components/numerical-guidance/molecule/metadata-dialog-menu/metadata-dialog-menu';
+import MetadataList from '@/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-list';
+import MetadataDialogMenu from '@/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-dialog-menu/metadata-dialog-menu';
 import userEvent from '@testing-library/user-event';
 
 describe('MetadataDialogMenu', () => {
@@ -18,7 +18,7 @@ describe('MetadataDialogMenu', () => {
     render(
       <SWRProviderWithoutCache>
         <MetadataDialogMenu />
-        <MetdataList />
+        <MetadataList />
       </SWRProviderWithoutCache>,
     );
     expect(screen.queryByRole('dialog')).toBeNull();
@@ -45,7 +45,7 @@ describe('MetadataDialogMenu', () => {
     render(
       <SWRProviderWithoutCache>
         <MetadataDialogMenu />
-        <MetdataList />
+        <MetadataList />
       </SWRProviderWithoutCache>,
     );
 
@@ -72,7 +72,7 @@ describe('MetadataDialogMenu', () => {
     render(
       <SWRProviderWithoutCache>
         <MetadataDialogMenu />
-        <MetdataList />
+        <MetadataList />
       </SWRProviderWithoutCache>,
     );
