+++ web/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-delete-dialog.tsx
