+++ web/app/ui/pages/workspace/dialog-container.tsx
@@ -1,6 +1,6 @@
 import IndicatorDialogMenu from '../../components/numerical-guidance/molecule/indicator-dialog-menu';
-import MetadataDeleteDialog from '../../components/numerical-guidance/molecule/metadata-delete-dialog';
-import MetadataDialogMenu from '../../components/numerical-guidance/molecule/metadata-dialog-menu/metadata-dialog-menu';
+import MetadataDeleteDialog from '../../components/numerical-guidance/indicator-board-metadata/metadata-delete-dialog';
+import MetadataDialogMenu from '../../components/numerical-guidance/indicator-board-metadata/metadata-dialog-menu/metadata-dialog-menu';
 
 export default function DialogContainer() {
   return (
