+++ web/app/ui/pages/workspace/metadata-list-container.tsx
@@ -1,5 +1,5 @@
 import { SWRConfig } from 'swr';
-import MetadataList from '../../components/numerical-guidance/molecule/metadata-list';
+import MetadataList from '../../components/numerical-guidance/indicator-board-metadata/metadata-list';
 import ClientDataSuspense from '../../components/view/atom/client-data-suspense';
 
 export default function MetadataListContainer() {
